+++ OffRoadPortal/Models/CreateCarDto.cs
@@ -25,7 +25,7 @@ public class CreateCarDto
 
     [Required]
     public short? Year { get; set; }
-    public short? Engine { get; set; }
+    public double? Engine { get; set; }
     public FuelType Fuel { get; set; }
     public string? Description { get; set; }
     public string? ImageUrl { get; set; }
