+++ OffRoadPortal/Models/UpdateCarDto.cs
@@ -16,7 +16,7 @@ public class UpdateCarDto
     public string? Mark { get; set; }
     public string? Model { get; set; }
     public short? Year { get; set; }
-    public short? Engine { get; set; }
+    public double? Engine { get; set; }
     public FuelType Fuel { get; set; }
     public string? Description { get; set; }
     public string? ImageUrl { get; set; }
